keyboard shortcuts for commenting
